cleanup and documentation

+++ app/src/main/java/com/example/bindingframedimensions/bindingframedimensions/Binding.kt
@@ -6,13 +6,13 @@ import android.view.View
 @BindingAdapter("android:layout_width")
 fun setLayoutWidth(view: View, width: Int) {
     val layoutParams = view.layoutParams
-    layoutParams.width = width.toInt()
+    layoutParams.width = width
     view.layoutParams = layoutParams
 }
 
 @BindingAdapter("android:layout_height")
 fun setLayoutHeight(view: View, height: Int) {
     val layoutParams = view.layoutParams
-    layoutParams.height = height.toInt()
+    layoutParams.height = height
     view.layoutParams = layoutParams
 }
